# feat(json): optimize format types encoding