```
name              old time/op    new time/op    delta
EncodeIP-4          82.3ns ± 5%    43.3ns ± 1%   -47.38%  (p=0.000 n=15+15)
EncodeDate-4         186ns ± 2%     148ns ± 2%   -20.13%  (p=0.000 n=15+15)
EncodeTime-4         162ns ± 5%     126ns ± 2%   -22.09%  (p=0.000 n=13+14)
EncodeDateTime-4     349ns ± 2%     307ns ± 4%   -12.09%  (p=0.000 n=15+14)

name              old alloc/op   new alloc/op   delta
EncodeIP-4           16.0B ± 0%      0.0B       -100.00%  (p=0.000 n=15+15)
EncodeDate-4          1.00 ± 0%      0.00       -100.00%  (p=0.000 n=15+15)
EncodeTime-4          1.00 ± 0%      0.00       -100.00%  (p=0.000 n=15+15)
EncodeDateTime-4      1.00 ± 0%      0.00       -100.00%  (p=0.000 n=15+15)

name              old allocs/op  new allocs/op  delta
EncodeIP-4            1.00 ± 0%      0.00       -100.00%  (p=0.000 n=15+15)
EncodeDate-4          1.00 ± 0%      0.00       -100.00%  (p=0.000 n=15+15)
EncodeTime-4          1.00 ± 0%      0.00       -100.00%  (p=0.000 n=15+15)
EncodeDateTime-4      1.00 ± 0%      0.00       -100.00%  (p=0.000 n=15+15)
```